requirements.txt
Django
gunicorn
python-decouple
django-sslserver
django-extensions
django-axes
requests
django-redis